Zero build warnings

+++ IlluminoEngine/src/Core/Buffer.h
@@ -85,8 +85,8 @@ namespace IlluminoEngine
 		}
 
 		inline const std::vector<BufferElement>& GetElements() const { return m_Elements; }
-		inline uint32_t GetCount() const { return m_Elements.size(); }
-		inline uint32_t GetStride() const { return m_Stride; }
+		inline size_t GetCount() const { return m_Elements.size(); }
+		inline size_t GetStride() const { return m_Stride; }
 
 		std::vector<BufferElement>::iterator begin() { return m_Elements.begin(); }
 		std::vector<BufferElement>::iterator end() { return m_Elements.end(); }
